Extracted the People component logic into a new usePeople custom hook
